phase 0: RTC modernization, docker rebuild, cleanup
- rtc: node:15 -> bun, index.js -> index.ts with types, npm -> bun
- auth2/Dockerfile: new multi-stage (bun dev + static deploy)
- auth/Dockerfile: node:14 -> bun
- docker-compose.yml: pipenv -> uv, npm -> bun, nodemon -> bun run dev,
  added ui service (auth2), legacy auth shifted to port 3001
- removed docker-compose-lite.yml, docker-compose-nginx.yml, custom.conf
- removed web10-deploy/ (stale GCP configs, skaffold k8s manifests)
