+++ auth/Dockerfile
@@ -1,11 +1,10 @@
-FROM --platform=linux/x86_64 node:14-alpine as dev
-WORKDIR auth
+FROM oven/bun:1-alpine as dev
+WORKDIR /auth
 COPY auth/package.json /auth/package.json
-RUN npm install
-RUN npm i -g serve
+RUN bun install
 COPY sdk /auth/public/sdk
 COPY auth/. /auth
 
 FROM dev as deploy
-RUN npm run build
-CMD npx react-inject-env set && PORT=80 serve -s build
+RUN bun run build
+CMD PORT=80 bunx serve -s build
