--- custom.conf
@@ -1,2 +1,0 @@
-client_max_body_size 64M;
-
