--- docker-compose-nginx.yml
@@ -1,19 +1,0 @@
-
-version: "3"
-
-services:
-  nginx-proxy:
-    image: jwilder/nginx-proxy
-    ports:
-      - "80:80"
-    volumes:
-      - /var/run/docker.sock:/tmp/docker.sock:ro
-      - ./custom.conf:/etc/nginx/conf.d/custom.conf:ro
-    restart: always
-    networks:
-      - all-spark-proxy
-
-networks:
-  all-spark-proxy:
-    external: true
-
