+++ rtc/package.json
@@ -1,7 +1,19 @@
 {
+  "name": "web10-rtc",
+  "type": "module",
+  "scripts": {
+    "dev": "bun --watch index.ts",
+    "build": "tsc",
+    "start": "bun run index.ts"
+  },
   "dependencies": {
-    "axios": "^0.26.1",
-    "jsonwebtoken": "^8.5.1",
+    "axios": "^1.6.0",
+    "jsonwebtoken": "^9.0.0",
     "peer": "^0.6.1"
+  },
+  "devDependencies": {
+    "typescript": "~5.8.2",
+    "@types/jsonwebtoken": "^9.0.0",
+    "@types/node": "^22.0.0"
   }
 }
