--- web10-deploy/README.md
@@ -1,7 +1,0 @@
-## web10 deployment.
-
-There are two options for deploying web10 :
-1. One using kubernetes / cloud platforms.
-2. One using regular docker-compose.
-
-Both deployment options use regular
