--- web10-deploy/mongo_migrate/.env_example
@@ -1,3 +1,0 @@
-# Define source and destination connection details
-sourceConnection="mongodb://username:password@sourceHost:sourcePort"
-destConnection="mongodb://username:password@destHost:destPort"
