--- web10-deploy/mongo_migrate/README.md
@@ -1,6 +1,0 @@
-## mongo migration utility.
-i used this to migrate my mongo data from atlas to my local mongo.
-copy the .env_example, rename it to .env
-set the .env variables to a mongo source database to a destination db.
-run ./mongo-migrate.sh
-done!
