--- web10-deploy/mongo_migrate/migration_dir/.gitignore
@@ -1,1 +1,0 @@
-*.json
