--- web10-deploy/skaffold/dev_example/kuberapi.yml
@@ -1,28 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: api-dev
-spec:
-  selector:
-    matchLabels:
-      app: api-dev
-  template:
-    metadata:
-      labels:
-        app: api-dev
-    spec:
-      containers:
-        - name: api-dev
-          image: gcr.io/pure-phalanx-344719/api-dev
-          ports:
-            - containerPort: 80
-          env:
-          - name: PROVIDER
-            value: "api.web10.dev"
-          - name: DB
-            value: "dev"
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
