--- web10-deploy/skaffold/dev_example/kuberauth.yml
@@ -1,28 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: auth-dev
-spec:
-  selector:
-    matchLabels:
-      app: auth-dev
-  template:
-    metadata:
-      labels:
-        app: auth-dev
-    spec:
-      containers:
-        - name: auth-dev
-          image: gcr.io/pure-phalanx-344719/auth-dev
-          ports:
-            - containerPort: 80
-          env:
-            - name: REACT_APP_DEFAULT_API
-              value: "api.web10.dev"
-            - name: PORT
-              value: "80"
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
