--- web10-deploy/skaffold/dev_example/kubercrm.yml
@@ -1,23 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: crm-dev
-spec:
-  selector:
-    matchLabels:
-      app: crm-dev
-  template:
-    metadata:
-      labels:
-        app: crm-dev
-    spec:
-      containers:
-        - name: crm-dev
-          image: gcr.io/pure-phalanx-344719/crm-dev
-          ports:
-            - containerPort: 80
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
