--- web10-deploy/skaffold/dev_example/skaffold.yaml
@@ -1,35 +1,0 @@
-apiVersion: skaffold/v2beta4
-kind: Config
-metadata:
-  name: web10
-build:
-  artifacts:
-    - image: gcr.io/pure-phalanx-344719/api-dev
-      context: ../../api
-      sync:
-        infer:
-          - './app/*'
-    - image: gcr.io/pure-phalanx-344719/auth-dev
-      context: ../..
-      docker:
-        dockerfile: auth/Dockerfile
-        target: dev
-      sync:
-        infer:
-          - './*'
-    - image: gcr.io/pure-phalanx-344719/crm-dev
-      context: ../..
-      sync:
-        infer:
-        - 'crm/*'
-        - 'sdk/*'
-      docker:
-        dockerfile: crm/Dockerfile
-deploy:
-  kubectl:
-    manifests:
-      - kuberapi.yml
-      - kuberauth.yml
-      - kubercrm.yml
-
-# crm and social don't need a dev, since it is front end code.
