--- web10-deploy/skaffold/run_example/kuberapi.yml
@@ -1,30 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: api
-spec:
-  selector:
-    matchLabels:
-      app: api
-  template:
-    metadata:
-      labels:
-        app: api
-    spec:
-      containers:
-        - name: api
-          image: gcr.io/pure-phalanx-344719/api
-          ports:
-            - containerPort: 80
-          env:
-            - name: PROVIDER
-              value: "api.web10.app"
-            - name: DB
-              value: "deploy"
-            - name: STRIPE_STATUS
-              value: "live"
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
