--- web10-deploy/skaffold/run_example/kuberauth.yml
@@ -1,32 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: auth
-spec:
-  selector:
-    matchLabels:
-      app: auth
-  template:
-    metadata:
-      labels:
-        app: auth
-    spec:
-      containers:
-        - name: auth
-          image: gcr.io/pure-phalanx-344719/auth
-          ports:
-            - containerPort: 80
-          env:
-            - name: REACT_APP_DEFAULT_API
-              value: "api.web10.app"
-            - name: REACT_APP_VERIFY_REQUIRED
-              value: true
-            - name: REACT_APP_PAY_REQUIRED
-              value: true
-            - name: PORT
-              value: "80"
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
