--- web10-deploy/skaffold/run_example/kubercrm.yml
@@ -1,23 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: crm
-spec:
-  selector:
-    matchLabels:
-      app: crm
-  template:
-    metadata:
-      labels:
-        app: crm
-    spec:
-      containers:
-        - name: crm
-          image: gcr.io/pure-phalanx-344719/crm
-          ports:
-            - containerPort: 80
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
