--- web10-deploy/skaffold/run_example/kuberhome.yml
@@ -1,23 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: home
-spec:
-  selector:
-    matchLabels:
-      app: home
-  template:
-    metadata:
-      labels:
-        app: home
-    spec:
-      containers:
-        - name: home
-          image: gcr.io/pure-phalanx-344719/home
-          ports:
-            - containerPort: 80
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
