--- web10-deploy/skaffold/run_example/kubermail.yml
@@ -1,23 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: mail
-spec:
-  selector:
-    matchLabels:
-      app: mail
-  template:
-    metadata:
-      labels:
-        app: mail
-    spec:
-      containers:
-        - name: mail
-          image: gcr.io/pure-phalanx-344719/mail
-          ports:
-            - containerPort: 80
-          resources:
-            requests:
-              memory: "512Mi"
-              cpu: "250m"
-            limits:
