--- web10-deploy/skaffold/run_example/skaffold.yaml
@@ -1,37 +1,0 @@
-apiVersion: skaffold/v2beta4
-kind: Config
-metadata:
-  name: web10
-build:
-  artifacts:
-    - image: gcr.io/pure-phalanx-344719/api
-      context: ../../api
-    - image: gcr.io/pure-phalanx-344719/rtc
-      context: ../../rtc
-    - image: gcr.io/pure-phalanx-344719/auth
-      context: ../..
-      docker:
-        dockerfile: auth/Dockerfile
-        target: deploy
-    - image: gcr.io/pure-phalanx-344719/crm
-      context: ../..
-      docker:
-        dockerfile: crm/Dockerfile
-    - image: gcr.io/pure-phalanx-344719/docs
-      context: ../../sdk
-    - image: gcr.io/pure-phalanx-344719/home
-      context: ../../home
-    - image: gcr.io/pure-phalanx-344719/mail
-      context: ../..
-      docker:
-        dockerfile: mail/Dockerfile
-deploy:
-  kubectl:
-    manifests:
-      - kuberapi.yml
-      - kuberauth.yml
-      - kuberrtc.yml
-      - kuberdocs.yml      
-      - kuberhome.yml
-      - kubercrm.yml
-      - kubermail.yml
